Warning cleanup

+++ wintendo/wintendoCore/apu.h
@@ -3,7 +3,7 @@
 #include "stdafx.h"
 #include "common.h"
 
-static const uint32_t		ApuSamplesPerSec	= CPU_HZ;
+static const uint32_t		ApuSamplesPerSec	= static_cast<uint32_t>( CPU_HZ );
 static constexpr uint32_t	ApuBufferMs			= static_cast<uint32_t>( 1000.0f / MinFPS );
 static constexpr uint32_t	ApuBufferSize		= static_cast<uint32_t>( ApuSamplesPerSec *  ( ApuBufferMs / 1000.0f ) );
