+++ wintendo/wintendoCore/common.h
@@ -30,9 +30,9 @@ using apuSeqCycle_t		= std::chrono::duration< uint64_t, std::ratio<ApuSequenceDi
 using scanCycle_t		= std::chrono::duration< uint64_t, std::ratio<PpuCyclesPerScanline * PpuClockDivide, MasterClockHz> >; // TODO: Verify
 using frameRate_t		= std::chrono::duration< double, std::ratio<1, FPS> >;
 
-static constexpr float CPU_HZ = static_cast<float>( chrono::duration_cast<cpuCycle_t>( chrono::seconds( 1 ) ).count() );
-static constexpr float APU_HZ = static_cast<float>( chrono::duration_cast<apuCycle_t>( chrono::seconds( 1 ) ).count() );
-static constexpr float PPU_HZ = static_cast<float>( chrono::duration_cast<ppuCycle_t>( chrono::seconds( 1 ) ).count() );
+static constexpr uint64_t CPU_HZ = chrono::duration_cast<cpuCycle_t>( chrono::seconds( 1 ) ).count();
+static constexpr uint64_t APU_HZ = chrono::duration_cast<apuCycle_t>( chrono::seconds( 1 ) ).count();
+static constexpr uint64_t PPU_HZ = chrono::duration_cast<ppuCycle_t>( chrono::seconds( 1 ) ).count();
 
 const uint32_t KB_1		= 1024;
 const uint32_t KB_2		= ( 2 * KB_1 );
